Initial net6.h
Has the definition of an IPv6 address and IPv6 header. It may make sense
to separate the v4 support from net.h (or to include this in net.h).

Cc: Kim Phillips <kim.phillips@freescale.com>
Reviewed-by: Kim Phillips <kim.phillips@freescale.com>
Signed-off-by: Chris Packham <chris.packham@alliedtelesis.co.nz>

Series-to: uboot, net, hannah@marvell.com
Series-prefix: RFC
Series-version: 3
Series-changes: 2
- use __be16/__be32
- add ipv6_addr_v4mapped and ipv6_addr_is_isatap inline functions
Series-changes: 3
- add reviewed-by from Kim Phillips

Cover-letter:
IPv6 support

This series so far covers the introduction of a IP6addr_t and
printing/parsing of addresses. I have a patch for testing these basic
things which I haven't included in this series yet. Is there a unittest
facility for u-boot I should be using?

A few open questions

1) Presumably the majority of the actual V6 code would be included by a
config option (CONFIG_IPV6). How far should I take that? Should the
vsprintf code be conditional?

2) Our current out of tree code parallels net.c and net.h. Should I
continue this for the final version or integrate it into net.[ch].

3) rxhand_f currently takes an IPaddr_t. I haven't looked at the usage
of this yet but to support V6 this may need to be a new union or a void
*.

END