net: tsec: enable promiscuous mode for IPv6
For IPv6 neighbor discovery to operate correctly we need to register for
various multicast groups (because some neighbor discovery packets are
sent to various multicast addresses).

Signed-off-by: Chris Packham <chris.packham@alliedtelesis.co.nz>

Commit-notes:
Currently the multicast support in the tsec driver is conditional on
CONFIG_MCAST_TFTP and tsec_mcast_addr, eth_mcast_join and
eth_device->mcast all disagree on what the prototype for the function to
add a multicast address should be. Instead of tackling this I've taken
the cowards way out and just enabled promiscuous mode which means the
device sees any packet regardless of the destination address.
END