net: ipv6 support
Adds basic support for IPv6. Neighbor discovery and ping6 are the only
things supported at the moment. Stub code exists for UDP but it is
untested and unlikely to work.

Signed-off-by: Chris Packham <chris.packham@alliedtelesis.co.nz>

Series-notes:
Now we have something functional. With this you can do something like
'setenv ipaddr6 3ffe::2' and 'ping6 3ffe::1' should work.
I seem to have a problem that when you send a ping6 for a non-existent
address that ends up stuck and the next non-ipv6 net operation tries to
resolve it. I suspect this is because the pending neighbor discovery
information isn't cleaned up properly.
I'm sending this out now because I'll be away at a conference next week
so probably won't have any time to work on this but I did want to say
"hey ping6 works". When I get back I'll start tackling tftp over ipv6.
END

---
This patch is bigger than I'd like it to be but I'm not sure how to
split it up and keep the parts build able. One obvious thing would be to
remove the UDP stub which I'll be updating anyway in order to support
TFTP over IPv6