net: Add support for Tigon3 based Ethernet NICs
This patch contains a port of the Linux Tigon3 driver to U-Boot.
Many non-bootloader type things (TSO, firmware, jumbo-frames, etc)
have been removed to keep the driver small and simple.

Series-prefix: RFC
Series-version: 1
Cover-letter:
net: Support for Broadcom Tigon 3 NICs

I should probably explain the people involved here. Joshua did most of
the porting work. I (Chris) picked it up to finish things off and to
do some more testing. I'll also be the one trying to get this accepted
upstream.

Because my employer paid for us to do the work I've signed with my work
email address. But because I'm already subscribed I'll send the patches
using my gmail account.
END

Commit-notes:
This code is functional. It's been tested using a P2041RDB with the
PCI-e NIC installed.

We've been a bit inconsistent with the porting. Some code we've removed.
Other code we've #if 0'd out. What is preferred? How well should this
compare to the equivalent Linux driver?

Also because of u-boot/linux differences (and because because the
original linux source is not clean) this isn't checkpatch.pl clean. We
could put effort into cleaning this up but that would make the driver
even less comparable to it's linux peer.
END

Signed-off-by: Joshua Scott <joshua.scott@alliedtelesis.co.nz>
Signed-off-by: Chris Packham <chris.packham@alliedtelesis.co.nz>