net: tsec: enable promiscuous mode
IPv6 neighbor discovery uses various multicast addresses to send the
request and receive the response. For neighbor discovery to work
properly in U-boot the Ethernet device needs to support joining/leaving
various multicast groups or it needs to support multicast/promiscuous
mode. For the sake of simplicity the latter approach has been taken.

Signed-off-by: Chris Packham <chris.packham@alliedtelesis.co.nz>

Commit-notes:
Drivers that support multicast reception have it enabled/disabled with
CONFIG_MCAST_TFTP. It wouldn't be too hard to create a separate
CONFIG_MCAST that is selected by enabling CONFIG_MCAST_TFTP or
CONFIG_NET6 but for now I want to concentrate on getting the rest of the
IPv6 code in good shape.
END