net: add definition of udp_hdr
UDP is the same over IPv4 as it is over other protocols (i.e. IPv6) add
a definition of just the UDP header independent of the IPv4 header that
it may or may not be combined with.

Signed-off-by: Chris Packham <judge.packham@gmail.com>
Commit-notes:
Ideally struct ip_udp_hdr would be defined as

  struct ip_udp_hdr {
          struct ip_hdr  ip;
          struct udp_hdr udp;
  };

Implementing this touches more code that I really want to at this
point. Some of the code that currently uses struct ip_udp_hdr could
probably just use struct ip_hdr instead but some care would need to be
taken to much such a change.
END