mtd: nand: pxa3xx_nand: Increase initial buffer size
The initial buffer is used for the initial commands used to detect
a flash device (STATUS, READID and PARAM).

ONFI param page is 256 bytes, and there are three redundant copies
to be read. JEDEC param page is 512 bytes, and there are also three
redundant copies to be read. Hence this buffer should be at least
512 x 3. This commits rounds the buffer size to 2048.

[ Linux commit c16340973fcb6461474a9f811f7f3ff2f946b24c ]

Cc: Ezequiel Garcia <ezequiel@vanguardiasur.com.ar>
Signed-off-by: Chris Packham <judge.packham@gmail.com>
Series-prefix: RFC
Series-version: 2
Cover-letter:
pxa3xx_nand updates
I'm looking into the NAND support for the db-88f6820-amc board.
There are a number of changes in the pxa3xx_nand driver in Linux that
are relevant (not specifically to this boards but to Armada boards in
general). Some of these changes are cleanups and some are actual bug
fixes.

I'd really appreciate some testing on this. This gives me enough so I
can mount and read files out of a ubifs volume created by Linux on the
DB-88F6820-AMC board but I don't have access to any other Armada boards
to test with.

Where applicable I'll Cc the author of the equivalent Linux patch. I've
had to tweak a few of them to make them fit u-boot (e.g. to handle
platform data) but they should still look pretty close to their Linux
counterparts.

Note this series is dependent on
http://lists.denx.de/pipermail/u-boot/2016-October/270967.html
END
Series-changes: 2
- Add reference to Linux commit sha1.
- Remove spurious change