arm: mvebu: move SYS_MVEBU_PLL_CLOCK to Kconfig
The main PLL frequency is 2GHz for Armada-XP and 1GHZ for Armada 375,
38x and 39x.

[ Linux commit ae142bd9976532aa5232ab0b00e621690d8bfe6a ]

Signed-off-by: Chris Packham <judge.packham@gmail.com>
Commit-notes:
See https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=ae142bd99765

I've set the value for ARMADA_3700 and ARMADA_8K to 2GHz. I'm not sure
if this is correct but it is the same value that these SoCs would have
had when the value was defined in soc.h.
END